seperate image usages into new service
