+++ kahuna/public/js/image/service.js
@@ -9,7 +9,6 @@ imageService.factory('imageService', [function() {
     function forImage(image) {
         return {
             usageRights: usageRights(image),
-            usages: usages(image),
             states: getStates(image)
         };
     }
@@ -18,36 +17,6 @@ imageService.factory('imageService', [function() {
         return {
             image: image,
             data: image.data.usageRights
-        };
-    }
-
-    function usages(image) {
-        function usageTitle(usage) {
-            const referenceType = usage.get('usageType') == 'print' ? 'indesign' : 'frontend';
-
-            const build = (usage, referenceType) => {
-                const reference = usage.get('references').find(u =>
-                    u.get('referenceType') == referenceType);
-
-                return reference ? reference.get('name') : 'No title found.';
-            };
-
-            return build(usage, referenceType);
-        }
-
-        const usagesList = Immutable.fromJS(image.data.usages).map(u =>
-            u.set('title', usageTitle(u)));
-
-        const groupedByState = usagesList.groupBy(u => u.get('status'));
-
-        const filterByUsageType = (usageType) =>
-            usagesList.filter(u => u.get('usageType') == usageType);
-
-        return {
-            usages: usagesList.toJS(),
-            groupedByState: groupedByState.toJS(),
-            hasPrintUsages: !filterByUsageType('print').isEmpty(),
-            hasWebUsages: !filterByUsageType('web').isEmpty()
         };
     }
